move get_router
